Submitted by Andrew Zabolotny: Patch #2995091: Fix gcc 4.4.1 warnings
Actually, I went a lot further, and made it completely -Wall -Wextra
clean for GCC 4.x.  (GCC 3.x users might want to use --disable-warnings
in configure.)


git-svn-id: svn://svn.code.sf.net/p/avarice/code/trunk/avarice@232 a7e36acc-7495-4a18-91ca-7b9a7881be12

+++ src/jtag2.h
@@ -152,7 +152,7 @@ class jtag2: public jtag
 	useDebugWire = useDW;
 	apply_nSRST = nsrst;
         is_xmega = xmega;
-	flashCachePageAddr = (unsigned long)-1;
+	flashCachePageAddr = (unsigned int)-1;
 	eepromCachePageAddr = (unsigned short)-1;
 	for (int i = 0; i < MAX_BREAKPOINTS2; i++)
 	  softBPcache[i].type = NONE;
